+++ src/assets/js/lumiere-highslide-options.js
@@ -1,7 +1,7 @@
 /* Settings for highslide */
 /* var lumiere_vars.imdb_path passed from wp_localize_script to class-movie.php */
 
-var url_highslide = lumiere_vars.imdb_path + "js/highslide/graphics/";
+var url_highslide = lumiere_vars.imdb_path + "assets/js/highslide/graphics/";
 var popup_border_colour = lumiere_vars.popup_border_colour;
 
 hs.allowWidthReduction = true;